requirements.txt
datasets==1.5.0
higher==0.2.1
huggingface-hub==0.0.8
torchvision==0.9.1
transformers==4.5.1
hydra-core
wandb
importlib-metadata
filelock
requests
protobuf
psutil
pyyaml
spacy
matplotlib
ipython
sentencepiece
